Remove href in favor of module prop
